Undoing commit 6df2ff077
The fix from that commit is not guaranteed to work because a hash
has no guaranteed order, so the stringification might turn out
differently. The intention of this fix was to make it possible
to unclaim items from the callback function that was passed to
the consume() method. From a FIFO perspective, this is a bit odd.
However, it is possible to call enqueue_item() to put the item in
the queue again.

The callback function is not supposed to interfere with consume()
in claiming / unclaiming and in fact it can't because the callback
does not have the "raw" item, only the enqueued data.