Two timestamps for requeued items
Until now there was only one timestamp, which was the creation timestamp.
In almost every case, items are created and put into the queue at the
same time. However, if an item failes and is requeued, these timestamps
are different. It can be very useful to have them both.

For normal cases items will have one timestamp "t" which represents
creation time and time it is put into the queue. (We used to have this so far).
On requeuing the item, the original timestamp "t" will be copied to
"t_created" and "t" will get the actual timestamp, so time of requeuing.

The method age() will now return the age() of being in the queue (which
is reset when requeuing) instead of age of the item since creation.